Custom Test
Previously, all unit test workflows were written to run exactly `richgo test
./...`.  Not all projects follow this convention and needed to customize this
behavior.  This change adds an optional configuration of the steps of a test
run.
