+++ octo/octo.go
@@ -24,14 +24,6 @@ import (
 	"sort"
 
 	"github.com/disiqueira/gotree"
-	"github.com/iancoleman/strcase"
-	"gopkg.in/yaml.v3"
-
-	"github.com/paketo-buildpacks/pipeline-builder/octo/actions"
-	"github.com/paketo-buildpacks/pipeline-builder/octo/actions/event"
-	"github.com/paketo-buildpacks/pipeline-builder/octo/dependabot"
-	"github.com/paketo-buildpacks/pipeline-builder/octo/labels"
-	"github.com/paketo-buildpacks/pipeline-builder/octo/release"
 )
 
 //go:generate statik -src . -include *.sh
@@ -173,229 +165,4 @@ func (Octo) Write(descriptor Descriptor, contributions []Contribution) error {
 
 	fmt.Println(t.Print())
 	return nil
-}
-
-type Contribution struct {
-	Path        string
-	Permissions os.FileMode
-	Structure   gotree.Tree
-	Content     []byte
-}
-
-func NewActionContribution(workflow actions.Workflow) (Contribution, error) {
-	var (
-		c   Contribution
-		err error
-	)
-	c.Path = filepath.Join(".github", "workflows", fmt.Sprintf("%s.yml", strcase.ToKebab(workflow.Name)))
-	c.Permissions = 0644
-
-	var t []event.Type
-	for k := range workflow.On {
-		t = append(t, k)
-	}
-
-	c.Structure = gotree.New(fmt.Sprintf("%s %s", c.Path, t))
-	for _, j := range workflow.Jobs {
-		c.Structure.Add(j.Name)
-	}
-
-	if c.Content, err = yaml.Marshal(workflow); err != nil {
-		return Contribution{}, fmt.Errorf("unable to marshal workflow\n%w", err)
-	}
-
-	return c, err
-}
-
-func NewDependabotContribution(dependabot dependabot.Dependabot) (Contribution, error) {
-	var (
-		c   Contribution
-		err error
-	)
-	c.Path = filepath.Join(".github", "dependabot.yml")
-	c.Permissions = 0644
-
-	c.Structure = gotree.New(c.Path)
-	for _, u := range dependabot.Updates {
-		c.Structure.Add(fmt.Sprintf("%s: %s", u.PackageEcosystem, u.Schedule.Interval))
-	}
-
-	if c.Content, err = yaml.Marshal(dependabot); err != nil {
-		return Contribution{}, fmt.Errorf("unable to marshal dependabot\n%w", err)
-	}
-
-	return c, err
-}
-
-func NewDockerCredentialActions(credentials []DockerCredentials) []actions.Step {
-	var s []actions.Step
-
-	for _, c := range credentials {
-		s = append(s, actions.Step{
-			Name: fmt.Sprintf("Docker login %s", c.Registry),
-			If:   "${{ github.event_name != 'pull_request' || ! github.event.pull_request.head.repo.fork }}",
-			Uses: "docker/login-action@v1",
-			With: map[string]interface{}{
-				"registry": c.Registry,
-				"username": c.Username,
-				"password": c.Password,
-			},
-		})
-	}
-
-	return s
-}
-
-func NewHttpCredentialActions(credentials []HTTPCredentials) []actions.Step {
-	var s []actions.Step
-
-	for _, c := range credentials {
-		s = append(s, actions.Step{
-			Name: fmt.Sprintf("HTTP login %s", c.Host),
-			If:   "${{ github.event_name != 'pull_request' || ! github.event.pull_request.head.repo.fork }}",
-			Run:  statikString("/update-netrc.sh"),
-			Env: map[string]string{
-				"HOST":     c.Host,
-				"USERNAME": c.Username,
-				"PASSWORD": c.Password,
-			},
-		})
-	}
-
-	return s
-}
-
-func NewDrafterContribution(drafter release.Drafter) (Contribution, error) {
-	var (
-		c   Contribution
-		err error
-	)
-	c.Path = filepath.Join(".github", "release-drafter.yml")
-	c.Permissions = 0644
-
-	c.Structure = gotree.New(c.Path)
-	for _, a := range drafter.Categories {
-		c.Structure.Add(fmt.Sprintf("%s: %s", a.Title, a.Labels))
-	}
-
-	if c.Content, err = yaml.Marshal(drafter); err != nil {
-		return Contribution{}, fmt.Errorf("unable to marshal release drafter\n%w", err)
-	}
-
-	return c, err
-}
-
-func NewLabelsContribution(labels []labels.Label) (Contribution, error) {
-	var (
-		c   Contribution
-		err error
-	)
-	c.Path = filepath.Join(".github", "labels.yml")
-	c.Permissions = 0644
-
-	c.Structure = gotree.New(c.Path)
-	for _, l := range labels {
-		c.Structure.Add(l.Name)
-	}
-
-	if c.Content, err = yaml.Marshal(labels); err != nil {
-		return Contribution{}, fmt.Errorf("unable to marshal labels\n%w", err)
-	}
-
-	return c, err
-}
-
-type Descriptor struct {
-	Path              string
-	CodeOwners        []CodeOwner
-	Builder           *Builder
-	Package           *Package
-	DockerCredentials []DockerCredentials `yaml:"docker_credentials"`
-	HttpCredentials   []HTTPCredentials   `yaml:"http_credentials"`
-	OfflinePackages   []OfflinePackage    `yaml:"offline_packages"`
-	Actions           []Action
-	Dependencies      []Dependency
-}
-
-type Action struct {
-	Source string
-	Target string
-}
-
-type Builder struct {
-	Repository string
-}
-
-type CodeOwner struct {
-	Path  string
-	Owner string
-}
-
-type DockerCredentials struct {
-	Registry string
-	Username string
-	Password string
-}
-
-type HTTPCredentials struct {
-	Host     string
-	Username string
-	Password string
-}
-
-type Dependency struct {
-	Name           string
-	Id             string
-	VersionPattern string `yaml:"version_pattern"`
-	Uses           string
-	With           map[string]interface{}
-}
-
-type OfflinePackage struct {
-	Source string
-	Target string
-}
-
-type Package struct {
-	Repository          string
-	IncludeDependencies bool `yaml:"include_dependencies"`
-	Register            bool
-	RegistryToken       string `yaml:"registry_token"`
-}
-
-func NewDescriptor(path string) (Descriptor, error) {
-	in, err := os.Open(path)
-	if err != nil {
-		return Descriptor{}, fmt.Errorf("unable to open %s\n%w", path, err)
-	}
-	defer in.Close()
-
-	var d Descriptor
-	if err := yaml.NewDecoder(in).Decode(&d); err != nil {
-		return Descriptor{}, fmt.Errorf("unable to decode descriptor from %s\n%w", path, err)
-	}
-
-	if d.Path == "" {
-		d.Path = ".."
-	}
-
-	if !filepath.IsAbs(d.Path) {
-		if d.Path, err = filepath.Abs(filepath.Join(filepath.Dir(path), d.Path)); err != nil {
-			return Descriptor{}, fmt.Errorf("unable to find absolute path\n%w", err)
-		}
-	}
-
-	for i, e := range d.Dependencies {
-		if e.Name == "" {
-			e.Name = e.Id
-			d.Dependencies[i] = e
-		}
-
-		if e.VersionPattern == "" {
-			e.VersionPattern = `[\d]+\.[\d]+\.[\d]+`
-			d.Dependencies[i] = e
-		}
-	}
-
-	return d, nil
 }
